backend/requirements.txt
asgiref==3.7.2
Django==4.2.4
django-cors-headers==4.2.0
djangorestframework==3.14.0
psycopg2-binary==2.9.7
Pygments==2.16.1
pytz==2023.3
sqlparse==0.4.4
